fix: rename authorizer models module to top-level name (avoid test collision)
